Do not use pooled connections
It is not clear whether pooled connections can be the cause of some
lockups. But since they limit the number of sockets used to reach the
same destination simultaneously, and since a socket is not freed if its
input stream is not explicitely closed (in a finally() clause ideally),
this may be a problem with the legacy code.

Use non-pooled connections (as was done before), but with proper
timeout handling.