Remove obsolete cache files
This fixes #609. Obsolete cache files will be removed at the
same time as the database entries.

Also, when a list is forgotten, the caches in the list are
not physically removed from the database. They are returned
to their previous status, and will be expired as needed.

At the same time, we add new indices on latitude and longitude
as they are used in map-related queries.

The first start will remove every repository located under
/sdcard/.cgeo which does not correspond to a cache file.
This should take care of all leftovers from previous versions.
The removal is done by a background thread to avoid blocking
the database for too long as the test is performed during the
database upgrade which runs in a transaction.