Prevent null check warnings
The use of CollectionsUtils.isNotEmpty() is not detected as meaning
that the checked value is not null. Those calls have been inlined.

Also, at some other places, we know that a value cannot be null
but the compiler has no means to check it by itself because
this information is not conveyed in the method signatures.