Implementation for #977: Import renamed waypoint GPX files
The new Method is based on the gpx file (not only the file name) and
checks all files in the directory for a matching filename.

Note: I updated (just implemented) a test method but couldn't already
test it. I'm working to get the tests running ... currently something
isn't configured right and i don't get it running.