Fix #1332: uppercase geocode built from name when importing GPX
When importing a GPX file, if no geocode is found matching the expected
format, one is built from the cache name. In several places, the geocode
is transformed into upper case, which is supposed to be a no-op. Use
uppercase when building a geocode from the cache name to ensure it is
indeed a no-op.