Partial fix: do not crash if a cache has no coordinates
It looks like from time to time some caches are sent with null
coordinates, at least when browsing the live map.

This change prevents a NullPointerException from being thrown and logs
this as an error in the log. This issue still needs to be addressed, as
the root cause of the problem is yet unknown.