Remove obsolete comment
We do not like to have variables initialized with a dummy value. It
prevents the flow path analysis from noticing a problem, as it would not
see that, in some cases, we have forgotten to initialize the variable.
Taping over potential mistakes, especially in a preventive way, is
definitely not a good idea.