Fix #1790: use correct freshness information for pictures
On some filesystems and Android versions, the "last modified" data on
files cannot be modified after the file has been initially created.
Since we use this information to indicate that a cached image has been
refreshed, this could fail on some devices.

Now, we use a degraded version of File#setLastModified if it returns
false. This goes through an intermediate copy, but is always much more
faster than getting the file from the network connection and saving it
on the local storage.