Reclaim bitmap memory in onStop() rather than in onDestroy()
onDestroy() might be called only when the system is low in main memory,
while we want to free the bitmap memory as soon as we do not need it
anymore. Since the images are cached on the local filesystem, redrawing
them after we come back from the image viewer does not take long.