Search for maps on multiple storages
Many devices are mounting /mnt/sdcard to internal flash memory. This memory is then
referred to as the ExternalStorage. These devices may still have the possibility to
insert an SDcard and that memory is hard to access. This modification will search all
mounted memory that is available for the user for map files.