Workaround for #2851
I inserted the include option for the new (temporary?) cache detail page and added an append command for the dd tag which is used instead of a p tag in the new cache page.
I know that the append is called twice in this case. Maybe someone can optimize it but it has shown no side effects at tests with both URLs.