Remove elevation handling
Geocaches do not contain useful elevation information. We used to get it
using Google API, but the information was not very useful because our
handling of altitude was inconsistent.

Also, the Google elevation API usage limits states that "the Elevation
API may only be used in conjunction with displaying results on a Google
map; using elevation data without displaying a map for which elevation
data was requested is prohibited".

This removes elevation handling completely, except in the settings to
avoid getting in the way of the settings rewrite. The elevation related
settings should be removed there instead when both works are merged.