fix #3521: GPS would stay on after a cache list refresh
The previous fix for #3521 was incomplete and only reduced the time
window during which the GPS could be turned on in a race condition
(4c5ef68f45a2b35537b89f857c1aa9fcb117ae83).

This one prevents the GPS from being turned on again when the
activity is paused (which includes a dead activity as well if
the user has left).