Remove onTerminate() in application
Not only it is not advised to derive from `Application` as this is often
deemed useless, but also the `onTerminate()` method is never called on
real devices.

Since it causes issues with our tests since it accesses ressources that
appear to have been cleaned up already, such as the application context,
it is best to remove it entirely.