fix #4441: wrong distance in lists for caches with modified coordinates
The coordinates used to get the cache location were taken from the .loc
data for premium members of geocaching.com and basic members with
captcha enabled. This has been changed to not use the data from the .loc
data when precise location is already available from the database or the
cache.