fix #4325: retrieve home location from geocaching.com
When the device does not have a location, the address from
geocaching.com (persisted accross sessions) will be used if it is
available instead of the last fallback (Paris).

However, the location will be retrieved only the second time the
application is launched, as we need an initial location before we can
get the home location from geocaching.com.