Shortened changelog
As usual for a new feature release the old changes are taken out of the changelog to keep it comprehensive.
I also had to remove the link to the github milestone as "Detailed list of changes" as this does no longer seem to be a permalink but an URL built with the milestone name which might change. Any idea for that?