Do not lazy load fields if the cache doesn't come from the database
When a cache is stored in the database, some fields are lazily loaded
from the database. This should only happen when the cache itself has
been loaded from the database, otherwise we might fill fields of a cache
acquired from the network with older version of the fields as stored in
the database.

Noticed during the work on #4576.