fix #4481: numerous images in description or log may lead to ANR
The forced redrawing of a text view caused by settings of text is a
lengthy operation. When we do it for every image, we might end up with
hundredths of operations waiting to do the same thing, with only the
first few ones being actually useful because images won't change
afterwards.

We now use a queue to ensure that consecutive redraw orders are done as
one operation.