fix #4881, #4658: wrong default settings
We must be more careful when changing defaults in the settings. Until
now we mostly just changed the default value in the Java getters. That
doesn't work, because on a fresh installation we first apply the XML
resource for the settings, therefore the defaults from the XML file are
taken.