Make "Found" status in GPX exports optional (#5835)
Add a checkbox to the export to GPX dialog to include/exclude the Found
status. When checked (default), the sym tag of the exported GPX will
contain the text "Geocache Found" for geocaches that were found. When
unchecked, the sym tag of the exported GPX will always show "Geocache"
and thus omits the word "Found".